Added tmux 1.9 ppa install

+++ tmux/install.sh
@@ -3,7 +3,13 @@ if ! hasCommand "tmux"; then
   if isOSX; then
     brew install tmux
   elif isLinux; then
-    sudo apt-get -y install tmux
+    if [ "$( echo "$(cat /etc/lsb-release | grep -o -m 1 "[0-9]*\.[0-9]*") 12.04" | awk -f ~/dotfiles/script/version.awk)" -eq 0 ]; then
+      sudo add-apt-repository -y ppa:pi-rho/dev
+      sudo apt-get update
+      sudo apt-get install -y tmux=1.9a-1~ppa1~p
+    else
+      sudo apt-get -y install tmux
+    fi
   fi
 fi
